Restore old, correct prev_real_wid behaviour.
Also a small lattice generation optimization - because of the
topological ordering of nodes, no nodes before dag->end are reachable
by definition, so we can immediately eliminate them.  In addition we
were not properly finding the end node - if </s> exists in a frame
before the end frame then we prefer it to all other nodes.